Make node sanitization optional

+++ install_agent.sh
@@ -90,6 +90,10 @@ rm /root/.ssh/known_hosts
 # remove mcollective and puppet logs
 rm -f /var/log/mcollective.log /var/log/puppet/*
 
-# sanitize the AMI (creating unique ssh host keys will be handled by the cloud-init package)
-find /root/.*history /home/*/.*history -exec rm -f {} \;
-find /root /home /etc -name "authorized_keys" -exec rm -f {} \;
+args="$*"
+
+if [[ ${args} =~ "--sanitize" ]]; then
+  # sanitize the AMI (creating unique ssh host keys will be handled by the cloud-init package)
+  find /root/.*history /home/*/.*history -exec rm -f {} \;
+  find /root /home /etc -name "authorized_keys" -exec rm -f {} \;
+fi;
